requirements.txt
altair
aniso8601
anyio
api
appnope
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
attrs
Babel
beautifulsoup4
bleach
blinker
cachetools
certifi
cffi
charset-normalizer
click
colorama
comm
contourpy
cors
cycler
debugpy
decorator
defusedxml
executing
fastjsonschema
filelock
Flask==3.0.0
Flask-Cors==4.0.0
flask-restplus==0.13.0
fonttools
fqdn
future
gevent
gitdb
GitPython
greenlet
gunicorn
idna
importlib-metadata
ipykernel
ipython
ipython-genutils
ipywidgets
isoduration
itsdangerous
jedi
Jinja2
joblib
json5
JsonForm
jsonpointer
jsonschema
jsonschema-specifications
JsonSir
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab-widgets
jupyterlab_server
kiwisolver
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
nbclient
nbconvert
nbformat
nest-asyncio
nltk
nose
notebook
notebook_shim
numpy
overrides
packaging
pandas
pandocfilters
parso
pexpect
Pillow
platformdirs
plotly
prometheus-client
prompt-toolkit
protobuf
psutil
ptyprocess
pure-eval
pyarrow
pycparser
pydeck
Pygments
pyparsing
PySocks
python-dateutil
python-dotenv
Python-EasyConfig
python-json-logger
pytz
PyYAML
pyzmq
qtconsole
QtPy
recommend
referencing
regex
requests
requests-file
Resource
rfc3339-validator
rfc3986-validator
rich
rpds-py
scikit-learn
scipy
seaborn
Send2Trash
six
smmap
sniffio
soupsieve
stack-data
streamlit
tenacity
terminado
threadpoolctl
tinycss2
tldextract
toml
toolz
tornado
tqdm
traitlets
types-python-dateutil
typing_extensions
tzdata
tzlocal
uri-template
urllib3
validators
wcwidth
webcolors
webencodings
websocket-client
Werkzeug
widgetsnbextension
zipp
zope.event
zope.interface
